Use gunicorn with 4 workers and preload AI models
This adds gunicorn to Pipfile/Pipfile.lock, rearranges things a bit in
v5_Gesture_Recognition_Serving.py so that not so much is gated on the
__main__ check, and adds a wsgi.py for the gunicorn entrypoint.

I verified v5_Gesture_Recognition_Serving.py still works locally both
when run directly and when run with gunicorn via:

    gunicorn --bind 0.0.0.0:5000 -w 4 wsgi

The presence of gunicorn in Pipfile and wsgi.py file should be enough
to get OpenShift to automatically use it for serving the app.

I chose 4 workers because it would be better to fully saturate the CPU
of each AI pod than to leave spare CPU on the table for each. We may
want to lower the maxScale value after this change.